Reimplement complaint issue banner with required flag conditions
