+++ cfgov/v1/jinja2tags/__init__.py
@@ -8,7 +8,9 @@ from jinja2.ext import Extension
 from v1.jinja2tags.datetimes import DatetimesExtension
 from v1.jinja2tags.fragment_cache import FragmentCacheExtension
 from v1.models import CFGOVRendition
-from v1.templatetags.complaint_issue_banner import complaint_issue_banner
+from v1.templatetags.complaint_banners import (
+    complaint_issue_banner,
+)
 from v1.templatetags.email_popup import email_popup
 from v1.templatetags.mega_menu import get_menu_items
 from v1.util import ref
